feat: user profile tab
