+++ src/social/features/user/Profile/components/index.tsx
@@ -4,3 +4,4 @@ export * from './Feed';
 export * from './ImageFeed';
 export * from './VideoFeed';
 export * from './Action';
+export * from './Menu';
