+++ src/social/screens/UserProfile/UserProfile.tsx
@@ -1,18 +1,14 @@
 import { useEffect, useState } from 'react';
-import AmityUserProfilePage from '../../features/user/Profile';
+import Profile from '../../features/user/Profile';
 import { RouteProp, useRoute, useNavigation } from '@react-navigation/native';
 import { RootStackParamList } from '../../../core/routes/RouteParamList';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 
 type UserProfileScreenProps = {
   defaultUserId?: string;
-  isShowBackButton?: boolean;
 };
 
-export function UserProfileScreen({
-  defaultUserId,
-  isShowBackButton,
-}: UserProfileScreenProps) {
+export function UserProfileScreen({ defaultUserId }: UserProfileScreenProps) {
   const route = useRoute<RouteProp<RootStackParamList, 'UserProfile'>>();
   const navigation =
     useNavigation<NativeStackNavigationProp<RootStackParamList>>();
@@ -33,10 +29,6 @@ export function UserProfileScreen({
   }, [defaultUserId, navigation, userIdFromRoute]);
 
   return (
-    <AmityUserProfilePage
-      userId={userId}
-      isFromComponent={!!defaultUserId && !userIdFromRoute}
-      isShowBackButton={isShowBackButton ? isShowBackButton : !defaultUserId}
-    />
+    <Profile userId={userId} inline={!!defaultUserId && !userIdFromRoute} />
   );
 }
